refactor: extract utils, uniforms composable, and address review feedback
