+++ src/renderer/glsl/useGLSLPreview.ts
@@ -3,155 +3,33 @@ import { computed, effectScope, onScopeDispose, ref, toValue, watch } from 'vue'
 
 import type { ComputedRef, EffectScope, MaybeRefOrGetter, Ref } from 'vue'
 import type { LGraphNode, NodeId } from '@/lib/litegraph/src/LGraphNode'
-import { SUBGRAPH_INPUT_ID } from '@/lib/litegraph/src/constants'
 import type { Subgraph } from '@/lib/litegraph/src/subgraph/Subgraph'
 import type { UUID } from '@/lib/litegraph/src/utils/uuid'
 import { useWorkflowStore } from '@/platform/workflow/management/stores/workflowStore'
 import { useNodeOutputStore } from '@/stores/nodeOutputStore'
 import { useWidgetValueStore } from '@/stores/widgetValueStore'
 
-import type { CurveData } from '@/components/curve/types'
-import { createInterpolator, isCurveData } from '@/components/curve/curveUtils'
+import { curveDataToFloatLUT } from '@/components/curve/curveUtils'
 import type { GLSLRendererConfig } from '@/renderer/glsl/useGLSLRenderer'
 import { useGLSLRenderer } from '@/renderer/glsl/useGLSLRenderer'
+import {
+  extractUniformSources,
+  getAutogrowLimits,
+  useGLSLUniforms
+} from '@/renderer/glsl/useGLSLUniforms'
 import {
   createSharedObjectUrl,
   releaseSharedObjectUrl
 } from '@/utils/objectUrlUtil'
 
-const GLSL_NODE_TYPE = 'GLSLShader'
-const DEBOUNCE_MS = 50
-const DEFAULT_SIZE = 512
-const MAX_PREVIEW_DIMENSION = 1024
-const CURVE_LUT_SIZE = 256
-
-interface AutogrowGroup {
-  max: number
-  min: number
-  prefix?: string
-}
-
-interface UniformSource {
-  nodeId: NodeId
-  widgetName: string
-}
-
-function getAutogrowLimits(node: LGraphNode): GLSLRendererConfig {
-  const defaults: GLSLRendererConfig = {
-    maxInputs: 5,
-    maxFloatUniforms: 20,
-    maxIntUniforms: 20,
-    maxBoolUniforms: 10,
-    maxCurves: 4
-  }
-
-  if (!('comfyDynamic' in node)) return defaults
-
-  const dynamic = node.comfyDynamic
-  if (
-    typeof dynamic !== 'object' ||
-    dynamic === null ||
-    !('autogrow' in dynamic)
-  )
-    return defaults
-
-  const groups = dynamic.autogrow as Record<string, AutogrowGroup> | undefined
-  if (!groups) return defaults
-
-  return {
-    maxInputs: groups['images']?.max ?? defaults.maxInputs,
-    maxFloatUniforms: groups['floats']?.max ?? defaults.maxFloatUniforms,
-    maxIntUniforms: groups['ints']?.max ?? defaults.maxIntUniforms,
-    maxBoolUniforms: groups['bools']?.max ?? defaults.maxBoolUniforms,
-    maxCurves: groups['curves']?.max ?? defaults.maxCurves
-  }
-}
-
-function normalizeDimension(value: unknown): number {
-  const parsed = Number(value)
-  if (!Number.isFinite(parsed) || parsed <= 0) return DEFAULT_SIZE
-  return parsed
-}
-
-function clampResolution(w: number, h: number): [number, number] {
-  const maxDim = Math.max(w, h)
-  if (maxDim <= MAX_PREVIEW_DIMENSION) return [w, h]
-  const scale = MAX_PREVIEW_DIMENSION / maxDim
-  return [Math.round(w * scale), Math.round(h * scale)]
-}
-
-function curveDataToLUT(curve: CurveData): Float32Array {
-  const lut = new Float32Array(CURVE_LUT_SIZE)
-  const interpolate = createInterpolator(curve.points, curve.interpolation)
-  for (let i = 0; i < CURVE_LUT_SIZE; i++) {
-    lut[i] = interpolate(i / (CURVE_LUT_SIZE - 1))
-  }
-  return lut
-}
-
-function getImageThroughSubgraphBoundary(
-  node: LGraphNode,
-  slot: number,
-  ownerSubgraphNode: LGraphNode
-): HTMLImageElement | undefined {
-  const graph = node.graph
-  if (!graph) return undefined
-
-  const input = node.inputs[slot]
-  if (input?.link == null) return undefined
-
-  const link = graph._links.get(input.link)
-  if (!link || link.origin_id !== SUBGRAPH_INPUT_ID) return undefined
-
-  const outerUpstream = ownerSubgraphNode.getInputNode(link.origin_slot)
-  if (!outerUpstream?.imgs?.length) return undefined
-
-  return outerUpstream.imgs[0]
-}
-
-function extractUniformSources(
-  glslNode: LGraphNode,
-  subgraph: Subgraph
-): {
-  floats: UniformSource[]
-  ints: UniformSource[]
-  bools: UniformSource[]
-  curves: UniformSource[]
-} {
-  const floats: UniformSource[] = []
-  const ints: UniformSource[] = []
-  const bools: UniformSource[] = []
-  const curves: UniformSource[] = []
-
-  if (!glslNode.inputs) return { floats, ints, bools, curves }
-
-  for (const input of glslNode.inputs) {
-    if (input.link == null) continue
-
-    const link = subgraph.getLink(input.link)
-    if (!link || link.origin_id === SUBGRAPH_INPUT_ID) continue
-
-    const sourceNode = subgraph.getNodeById(link.origin_id)
-    if (!sourceNode?.widgets?.[0]) continue
-
-    const inputName = input.name ?? ''
-    const dotIndex = inputName.indexOf('.')
-    if (dotIndex === -1) continue
-
-    const prefix = inputName.slice(0, dotIndex)
-    const source: UniformSource = {
-      nodeId: sourceNode.id as NodeId,
-      widgetName: sourceNode.widgets[0].name
-    }
-
-    if (prefix === 'floats') floats.push(source)
-    else if (prefix === 'ints') ints.push(source)
-    else if (prefix === 'bools') bools.push(source)
-    else if (prefix === 'curves') curves.push(source)
-  }
-
-  return { floats, ints, bools, curves }
-}
+import {
+  clampResolution,
+  DEBOUNCE_MS,
+  DEFAULT_SIZE,
+  getImageThroughSubgraphBoundary,
+  GLSL_NODE_TYPE,
+  normalizeDimension
+} from '@/renderer/glsl/glslPreviewUtils'
 
 /**
  * Two-tier composable for GLSL live preview.
@@ -251,14 +129,14 @@ function createInnerPreview(
   let rendererReady = false
   let renderRequestId = 0
 
-  const innerGLSLNode = computed(() => {
+  const innerGLSLNode = (() => {
     const node = nodeRef.value
     if (!node?.isSubgraphNode()) return null
     const subgraph = node.subgraph as Subgraph | undefined
     return subgraph?.nodes.find((n) => n.type === GLSL_NODE_TYPE) ?? null
-  })
+  })()
 
-  const ownerSubgraphNode = computed(() => {
+  const ownerSubgraphNode = (() => {
     const node = nodeRef.value
     const graph = node?.graph
     if (!graph) return null
@@ -266,11 +144,11 @@ function createInnerPreview(
     if (!rootGraph || graph === rootGraph) return null
 
     return (
-      rootGraph._nodes.find(
+      rootGraph._nodes?.find(
         (n) => n.isSubgraphNode() && n.subgraph === graph
       ) ?? null
     )
-  })
+  })()
 
   const graphId = computed(
     () => nodeRef.value?.graph?.rootGraph?.id as UUID | undefined
@@ -287,7 +165,7 @@ function createInnerPreview(
     const locatorId = nodeToNodeLocatorId(node)
     if (outputs[locatorId]?.images?.length) return true
 
-    const inner = innerGLSLNode.value
+    const inner = innerGLSLNode
     if (inner) {
       const innerLocatorId = nodeToNodeLocatorId(inner)
       if (outputs[innerLocatorId]?.images?.length) return true
@@ -296,13 +174,13 @@ function createInnerPreview(
     return false
   })
 
-  const isActive = computed(
+  const shouldRender = computed(
     () =>
       (isGLSLNode.value || isGLSLSubgraphNode.value) && hasExecutionOutput.value
   )
 
   watch(
-    isActive,
+    shouldRender,
     (v) => {
       isActiveOut.value = v
     },
@@ -321,7 +199,7 @@ function createInnerPreview(
         | undefined
     }
 
-    const inner = innerGLSLNode.value
+    const inner = innerGLSLNode
     if (inner) {
       return widgetValueStore.getWidget(
         gId,
@@ -334,7 +212,7 @@ function createInnerPreview(
   })
 
   const rendererConfig = computed(() => {
-    const inner = innerGLSLNode.value
+    const inner = innerGLSLNode
     if (inner) return getAutogrowLimits(inner)
 
     const node = nodeRef.value
@@ -351,169 +229,18 @@ function createInnerPreview(
 
   const uniformSources = computed(() => {
     const node = nodeRef.value
-    const inner = innerGLSLNode.value
+    const inner = innerGLSLNode
     if (!node?.isSubgraphNode() || !inner) return null
     return extractUniformSources(inner, node.subgraph as Subgraph)
   })
 
-  function collectUniformValues(
-    subgraphSources: UniformSource[] | undefined,
-    groupName: string,
-    uniformPrefix: string,
-    maxCount: number
-  ): number[] {
-    const gId = graphId.value
-    if (!gId) return []
-
-    if (subgraphSources) {
-      return subgraphSources.map(({ nodeId: nId, widgetName }) => {
-        const widget = widgetValueStore.getWidget(gId, nId, widgetName)
-        return Number(widget?.value ?? 0) || 0
-      })
-    }
-
-    const nId = nodeId.value
-    const node = nodeRef.value
-    if (nId == null || !node) return []
-
-    const values: number[] = []
-    for (let i = 0; i < maxCount; i++) {
-      const inputName = `${groupName}.${uniformPrefix}${i}`
-      const widget = widgetValueStore.getWidget(gId, nId, inputName)
-      if (widget !== undefined) {
-        values.push(Number(widget.value) || 0)
-        continue
-      }
-
-      const slot = node.inputs?.findIndex((inp) => inp.name === inputName)
-      if (slot == null || slot < 0) break
-
-      const upstreamNode = node.getInputNode(slot)
-      if (!upstreamNode) break
-      const upstreamWidgets = widgetValueStore.getNodeWidgets(
-        gId,
-        upstreamNode.id as NodeId
-      )
-      if (upstreamWidgets.length === 0) break
-      values.push(Number(upstreamWidgets[0].value) || 0)
-    }
-    return values
-  }
-
-  const floatValues = computed(() =>
-    collectUniformValues(
-      uniformSources.value?.floats,
-      'floats',
-      'u_float',
-      rendererConfig.value.maxFloatUniforms
-    )
+  const { floatValues, intValues, boolValues, curveValues } = useGLSLUniforms(
+    graphId,
+    nodeId,
+    nodeRef,
+    uniformSources,
+    rendererConfig
   )
-
-  const intValues = computed(() =>
-    collectUniformValues(
-      uniformSources.value?.ints,
-      'ints',
-      'u_int',
-      rendererConfig.value.maxIntUniforms
-    )
-  )
-
-  function collectBoolValues(
-    subgraphSources: UniformSource[] | undefined,
-    maxCount: number
-  ): boolean[] {
-    const gId = graphId.value
-    if (!gId) return []
-
-    if (subgraphSources) {
-      return subgraphSources.map(({ nodeId: nId, widgetName }) => {
-        const widget = widgetValueStore.getWidget(gId, nId, widgetName)
-        return Boolean(widget?.value ?? false)
-      })
-    }
-
-    const nId = nodeId.value
-    const node = nodeRef.value
-    if (nId == null || !node) return []
-
-    const values: boolean[] = []
-    for (let i = 0; i < maxCount; i++) {
-      const inputName = `bools.u_bool${i}`
-      const widget = widgetValueStore.getWidget(gId, nId, inputName)
-      if (widget !== undefined) {
-        values.push(Boolean(widget.value))
-        continue
-      }
-
-      const slot = node.inputs?.findIndex((inp) => inp.name === inputName)
-      if (slot == null || slot < 0) break
-
-      const upstreamNode = node.getInputNode(slot)
-      if (!upstreamNode) break
-      const upstreamWidgets = widgetValueStore.getNodeWidgets(
-        gId,
-        upstreamNode.id as NodeId
-      )
-      if (upstreamWidgets.length === 0) break
-      values.push(Boolean(upstreamWidgets[0].value))
-    }
-    return values
-  }
-
-  const boolValues = computed(() =>
-    collectBoolValues(
-      uniformSources.value?.bools,
-      rendererConfig.value.maxBoolUniforms
-    )
-  )
-
-  const curveValues = computed((): CurveData[] => {
-    const gId = graphId.value
-    if (!gId) return []
-
-    const sources = uniformSources.value?.curves
-    if (sources && sources.length > 0) {
-      return sources
-        .map(({ nodeId: nId, widgetName }) => {
-          const widget = widgetValueStore.getWidget(gId, nId, widgetName)
-          return widget && isCurveData(widget.value)
-            ? (widget.value as CurveData)
-            : null
-        })
-        .filter((v): v is CurveData => v !== null)
-    }
-
-    const node = nodeRef.value
-    const nId = nodeId.value
-    if (nId == null || !node?.inputs) return []
-
-    const values: CurveData[] = []
-    const max = rendererConfig.value.maxCurves
-    for (let i = 0; i < max; i++) {
-      const inputName = `curves.u_curve${i}`
-
-      const widget = widgetValueStore.getWidget(gId, nId, inputName)
-      if (widget && isCurveData(widget.value)) {
-        values.push(widget.value as CurveData)
-        continue
-      }
-
-      const slot = node.inputs.findIndex((inp) => inp.name === inputName)
-      if (slot < 0) break
-
-      const upstreamNode = node.getInputNode(slot)
-      if (!upstreamNode) break
-
-      const upstreamWidgets = widgetValueStore.getNodeWidgets(
-        gId,
-        upstreamNode.id as NodeId
-      )
-      const curveWidget = upstreamWidgets.find((w) => isCurveData(w.value))
-      if (!curveWidget) break
-      values.push(curveWidget.value as CurveData)
-    }
-    return values
-  })
 
   function loadInputImages(): void {
     const node = nodeRef.value
@@ -544,7 +271,7 @@ function createInnerPreview(
         continue
       }
 
-      const owner = ownerSubgraphNode.value
+      const owner = ownerSubgraphNode
       if (owner) {
         const img = getImageThroughSubgraphBoundary(node, slot, owner)
         if (img) {
@@ -586,7 +313,7 @@ function createInnerPreview(
         )
       }
 
-      const owner = ownerSubgraphNode.value
+      const owner = ownerSubgraphNode
       if (owner) {
         const img = getImageThroughSubgraphBoundary(node, slot, owner)
         if (img) {
@@ -650,7 +377,7 @@ function createInnerPreview(
   async function renderPreview(): Promise<void> {
     const requestId = ++renderRequestId
     const source = shaderSource.value
-    if (!source || !isActive.value) return
+    if (!source || !shouldRender.value) return
 
     const r = ensureRenderer()
 
@@ -687,7 +414,7 @@ function createInnerPreview(
       }
       const curves = curveValues.value
       for (let i = 0; i < curves.length; i++) {
-        r.bindCurveTexture(i, curveDataToLUT(curves[i]))
+        r.bindCurveTexture(i, curveDataToFloatLUT(curves[i]))
       }
 
       r.render()
@@ -695,19 +422,20 @@ function createInnerPreview(
       const blob = await r.toBlob()
       if (requestId !== renderRequestId || disposed) return
       const blobUrl = createSharedObjectUrl(blob)
-
-      const inner = innerGLSLNode.value
-      if (inner) {
-        const innerLocatorId = nodeToNodeLocatorId(inner)
-        nodeOutputStore.setNodePreviewsByLocatorId(innerLocatorId, [blobUrl])
-      } else {
-        const nId = nodeId.value
-        if (nId != null) {
-          nodeOutputStore.setNodePreviewsByNodeId(nId, [blobUrl])
+      try {
+        const inner = innerGLSLNode
+        if (inner) {
+          const innerLocatorId = nodeToNodeLocatorId(inner)
+          nodeOutputStore.setNodePreviewsByLocatorId(innerLocatorId, [blobUrl])
+        } else {
+          const nId = nodeId.value
+          if (nId != null) {
+            nodeOutputStore.setNodePreviewsByNodeId(nId, [blobUrl])
+          }
         }
+      } finally {
+        releaseSharedObjectUrl(blobUrl)
       }
-
-      releaseSharedObjectUrl(blobUrl)
     } catch (error) {
       if (requestId !== renderRequestId) return
       lastError.value =
@@ -720,7 +448,7 @@ function createInnerPreview(
   }, DEBOUNCE_MS)
 
   watch(
-    isActive,
+    shouldRender,
     (active) => {
       if (isGLSLNode.value) {
         const node = nodeRef.value
@@ -740,13 +468,13 @@ function createInnerPreview(
         curveValues.value
       ] as const,
     () => {
-      if (isActive.value) debouncedRender()
+      if (shouldRender.value) debouncedRender()
     },
     { deep: true }
   )
 
   watch(shaderSource, () => {
-    if (isActive.value) debouncedRender()
+    if (shouldRender.value) debouncedRender()
   })
 
   // Return dispose function for the inner tier
@@ -757,7 +485,7 @@ function createInnerPreview(
     renderer = null
 
     // Revoke preview blob URLs to avoid memory leaks
-    const inner = innerGLSLNode.value
+    const inner = innerGLSLNode
     if (inner) {
       const locatorId = nodeToNodeLocatorId(inner)
       nodeOutputStore.revokePreviewsByLocatorId(locatorId)
